requirements.txt
Flask
SQLAlchemy
psycopg2-binary
alembic
python-dotenv
pre-commit
flake8
flake8-import-order
PyJWT
loguru
